Add frontend prod dockerfile
- Make various changes to solve issues at build time
