+++ frontend/package.json
@@ -17,9 +17,8 @@
     "@mui/joy": "^5.0.0-beta.20",
     "@tanstack/react-query": "^4.35.3",
     "@testing-library/user-event": "^14.5.1",
-    "@types/node": "20.6.0",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
+    "@types/react": "18.2.23",
+    "@types/react-dom": "18.2.8",
     "date-fns": "^2.30.0",
     "eslint": "8.49.0",
     "eslint-config-next": "^14.0.4",
