+++ frontend/src/app/auth/error/page.tsx
@@ -1,14 +1,28 @@
 'use client'
+import { useSearchParams } from 'next/navigation'
 
-type ErrorProps = {
-    error: Error;
-}
-export default function Error({error}: ErrorProps) {
+export default function Error(){
+    const searchParams = useSearchParams()
+    const errorType = searchParams.get('error') ?? "1";
+    let message = "";
+
+    switch(errorType){
+        case "Configuration":
+            message = "There is a problem with the server configuration"
+            break;
+        case "AccessDenied":
+            message = "Access denied. You do not have permission to use the application."
+            break;
+        default:
+            message = "That was unexpected. Please contact an admin for assistance."
+    }
 
     return (
         <>
             <h1>Something went wrong</h1>
-            <p>{error?.message || ""}</p>
+            <p>
+                {message}
+            </p>
         </>
     )
 }
